# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FarmShield AI",
    "description": "\n**Hyperlocal climate risk infrastructure for smallholder farmers in Kenya.**\n\nFarmShield turns weather-station readings (JKUAT Conduit station) + crop + growth stage + location into:\n\n* a **Farm Risk Score** - drought, flood, heat-stress and crop-health sub-scores (LOW / MEDIUM / HIGH) and an overall 0-100 score, each backed by human-readable *reasons*;\n* **stage-aware advice** for the farmer in English and Kiswahili, delivered by SMS;\n* **parametric insurance trigger signals** insurers and SACCOs can act on.\n\nScoring is deterministic and rule-based (FAO-56 crop water requirements, KALRO crop guides) so every number is auditable.\nPartners authenticate with an `X-API-Key` header on the `/api/v1` routes.\n",
    "contact": {
      "name": "FarmShield AI - Hack the Weather 2026",
      "url": "https://github.com/Millie-source/FarmShield-AI"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/farms/{farm_id}/assess": {
      "post": {
        "tags": [
          "risk"
        ],
        "summary": "Run the risk engine for a farm now",
        "description": "Pulls the last 30 days of readings for the farm location, derives the growth stage, scores drought / flood / heat / crop health, evaluates the default drought insurance trigger and generates advice. The result is stored and returned.",
        "operationId": "assess_farm_farms__farm_id__assess_post",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          },
          {
            "name": "scenario",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "normal",
                    "dry_spell",
                    "heavy_rain"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Force a mock scenario for this assessment only",
              "title": "Scenario"
            },
            "description": "Force a mock scenario for this assessment only"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/farms/{farm_id}/risk": {
      "get": {
        "tags": [
          "risk"
        ],
        "summary": "Latest Farm Risk Score",
        "description": "Returns the most recent stored assessment; runs one if the farm has never been assessed.",
        "operationId": "get_risk_farms__farm_id__risk_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/farms/{farm_id}/risk/history": {
      "get": {
        "tags": [
          "risk"
        ],
        "summary": "Assessment history (newest first)",
        "operationId": "risk_history_farms__farm_id__risk_history_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiskHistoryItem"
                  },
                  "title": "Response Risk History Farms  Farm Id  Risk History Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenario": {
      "get": {
        "tags": [
          "demo"
        ],
        "summary": "Active mock weather scenario",
        "operationId": "get_scenario_scenario_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioOut"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "demo"
        ],
        "summary": "Switch the mock weather scenario (live demo)",
        "description": "Flips the replayed weather between normal / dry_spell / heavy_rain and, by default, re-assesses every farm so scores and SMS alerts change immediately.",
        "operationId": "set_scenario_scenario_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioSwitchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/risk/{farm_id}": {
      "get": {
        "tags": [
          "partners"
        ],
        "summary": "Dynamic risk score for one farm",
        "description": "Returns the latest Farm Risk Score: overall 0-100 + label, four sub-scores with reasons, growth stage, the default drought insurance trigger, advice text, timestamp and data sources. Set `fresh=true` to force a new assessment against current weather.",
        "operationId": "partner_risk_api_v1_risk__farm_id__get",
        "security": [
          {
            "PartnerApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          },
          {
            "name": "fresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Re-run the engine now instead of returning the stored latest assessment",
              "default": false,
              "title": "Fresh"
            },
            "description": "Re-run the engine now instead of returning the stored latest assessment"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-API-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdviceOut": {
        "properties": {
          "en": {
            "type": "string",
            "title": "En",
            "description": "Farmer-friendly advice in English"
          },
          "sw": {
            "type": "string",
            "title": "Sw",
            "description": "Ushauri kwa Kiswahili"
          },
          "source": {
            "type": "string",
            "enum": [
              "gemini",
              "fallback"
            ],
            "title": "Source",
            "description": "gemini when the LLM produced it, fallback for rule-based text"
          },
          "sms_en": {
            "type": "string",
            "title": "Sms En",
            "description": "<=160 char SMS version (EN)"
          },
          "sms_sw": {
            "type": "string",
            "title": "Sms Sw",
            "description": "<=160 char SMS version (SW)"
          }
        },
        "type": "object",
        "required": [
          "en",
          "sw",
          "source",
          "sms_en",
          "sms_sw"
        ],
        "title": "AdviceOut"
      },
...
      "ApiError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ApiError",
        "example": {
          "detail": "Farm 42 not found"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InsuranceTriggerOut": {
        "properties": {
          "triggered": {
            "type": "boolean",
            "title": "Triggered"
          },
          "rule": {
            "type": "string",
            "title": "Rule",
            "description": "Rule id: drought_rainfall_deficit | excess_rainfall | heat_days"
          },
          "evidence": {
            "type": "object",
            "title": "Evidence",
            "description": "Observed values behind the decision, for audit"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Data completeness x margin from threshold"
          },
          "policy": {
            "type": "object",
            "title": "Policy"
          }
        },
        "type": "object",
        "required": [
          "triggered",
          "rule",
          "evidence",
          "confidence",
          "policy"
        ],
        "title": "InsuranceTriggerOut",
        "example": {
          "confidence": 1.0,
          "evidence": {
            "deficit_mm": 29.6,
            "dry_days": 21,
            "rainfall_total_mm": 0.4,
            "stage": "flowering",
            "stage_is_critical": true,
            "threshold_mm": 30,
            "window_days": 21
          },
          "policy": {
            "critical_stages_only": false,
            "rainfall_threshold_mm": 30,
            "type": "drought",
            "window_days": 21
          },
          "rule": "drought_rainfall_deficit",
          "triggered": true
        }
      },
      "OverallOut": {
        "properties": {
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "level": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "title": "Level"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "\"LOW CLIMATE RISK\" | \"MODERATE CLIMATE RISK\" | \"HIGH CLIMATE RISK\""
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Weights",
            "description": "Stage-dependent weights applied to the sub-scores"
          }
        },
        "type": "object",
        "required": [
          "score",
          "level",
          "label",
          "weights"
        ],
        "title": "OverallOut"
      },
...
      "RiskHistoryItem": {
        "properties": {
          "assessment_id": {
            "type": "integer",
            "title": "Assessment Id"
          },
          "assessed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assessed At"
          },
          "scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "overall_score": {
            "type": "integer",
            "title": "Overall Score"
          },
          "overall_level": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "title": "Overall Level"
          },
          "drought": {
            "type": "integer",
            "title": "Drought"
          },
          "flood": {
            "type": "integer",
            "title": "Flood"
          },
          "heat": {
            "type": "integer",
            "title": "Heat"
          },
          "crop_health": {
            "type": "integer",
            "title": "Crop Health"
          },
          "insurance_triggered": {
            "type": "boolean",
            "title": "Insurance Triggered"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "assessed_at",
          "scenario",
          "stage",
          "overall_score",
          "overall_level",
          "drought",
          "flood",
          "heat",
          "crop_health",
          "insurance_triggered"
        ],
        "title": "RiskHistoryItem"
      },
      "RiskOut": {
        "properties": {
          "assessment_id": {
            "type": "integer",
            "title": "Assessment Id"
          },
          "farm_id": {
            "type": "integer",
            "title": "Farm Id"
          },
          "farm_name": {
            "type": "string",
            "title": "Farm Name"
          },
          "crop": {
            "type": "string",
            "title": "Crop"
          },
          "stage": {
            "$ref": "#/components/schemas/StageOut"
          },
          "overall": {
            "$ref": "#/components/schemas/OverallOut"
          },
          "sub_scores": {
            "$ref": "#/components/schemas/SubScoresOut"
          },
          "insurance_trigger": {
            "$ref": "#/components/schemas/InsuranceTriggerOut"
          },
          "advice": {
            "$ref": "#/components/schemas/AdviceOut"
          },
          "assessed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assessed At"
          },
          "scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario",
            "description": "Mock scenario the readings came from (null for live data)"
          },
          "data_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data Sources"
          },
          "readings_used": {
            "type": "integer",
            "title": "Readings Used"
          },
          "window_days": {
            "type": "integer",
            "title": "Window Days"
          },
          "ndvi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ndvi",
            "description": "Satellite NDVI if a provider supplied one"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "farm_id",
          "farm_name",
          "crop",
          "stage",
          "overall",
          "sub_scores",
          "insurance_trigger",
          "advice",
          "assessed_at",
          "data_sources",
          "readings_used",
          "window_days"
        ],
        "title": "RiskOut",
        "description": "Full Farm Risk Score: the core product of FarmShield.",
        "example": {
          "advice": {
            "en": "Your maize is flowering and has had almost no rain for 3 weeks. Irrigate within 24-48 hours if you can. Do not apply fertiliser until rain returns.",
            "sms_en": "FarmShield: Kamau Maize Plot risk HIGH (72/100). Maize flowering, no rain 3 wks. Irrigate in 24-48h. Hold fertiliser.",
            "sms_sw": "FarmShield: Kamau Maize Plot hatari JUU (72/100). Mahindi yanatoa maua, hakuna mvua wiki 3. Mwagilia ndani ya saa 24-48.",
            "source": "fallback",
            "sw": "Mahindi yako yanatoa maua na hayajapata mvua kwa wiki 3. Mwagilia ndani ya saa 24-48 ukiweza. Usiweke mbolea hadi mvua irudi."
          },
          "assessed_at": "2026-09-03T08:00:00Z",
          "assessment_id": 12,
          "crop": "maize",
          "data_sources": [
            "mock:dry_spell"
          ],
          "farm_id": 1,
          "farm_name": "Kamau Maize Plot",
          "insurance_trigger": {
            "confidence": 1.0,
            "evidence": {
              "rainfall_total_mm": 0.4,
              "threshold_mm": 30,
              "window_days": 21
            },
            "policy": {
              "rainfall_threshold_mm": 30,
              "type": "drought",
              "window_days": 21
            },
            "rule": "drought_rainfall_deficit",
            "triggered": true
          },
          "overall": {
            "label": "HIGH CLIMATE RISK",
            "level": "HIGH",
            "score": 72,
            "weights": {
              "crop_health": 0.2,
              "drought": 0.4,
              "flood": 0.1,
              "heat": 0.3
            }
          },
          "readings_used": 30,
          "scenario": "dry_spell",
          "stage": {
            "day_after_planting": 62,
            "day_in_stage": 12,
            "is_critical": true,
            "name": "flowering",
            "progress": 0.496,
            "sensitivity": 1.0,
            "stage_length_days": 25,
            "water_need_mm_week": 37.8
          },
          "sub_scores": {
            "crop_health": {
              "label": "POOR",
              "level": "HIGH",
              "reasons": [
                "No satellite data: health inferred from weather stress over 30 days"
              ],
              "score": 61
            },
            "drought": {
              "level": "HIGH",
              "reasons": [
                "Only 0 mm rain in the last 7 days vs 38 mm/week needed at flowering"
              ],
              "score": 81
            },
            "flood": {
              "level": "LOW",
              "reasons": [
                "0 mm in the last 72 h, below the 90 mm flood threshold"
              ],
              "score": 0
            },
            "heat": {
              "level": "HIGH",
              "reasons": [
                "6 of the last 7 days exceeded 32°C (peak 33.9°C) during flowering"
              ],
              "score": 68
            }
          },
          "window_days": 30
        }
      },
      "RiskSummary": {
        "properties": {
          "assessment_id": {
            "type": "integer",
            "title": "Assessment Id"
          },
          "farm_id": {
            "type": "integer",
            "title": "Farm Id"
          },
          "assessed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assessed At"
          },
          "overall_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Overall Score"
          },
          "overall_level": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "title": "Overall Level"
          },
          "overall_label": {
            "type": "string",
            "title": "Overall Label",
            "description": "e.g. \"HIGH CLIMATE RISK\""
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "insurance_triggered": {
            "type": "boolean",
            "title": "Insurance Triggered"
          },
          "scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "farm_id",
          "assessed_at",
          "overall_score",
          "overall_level",
          "overall_label",
          "stage",
          "insurance_triggered"
        ],
        "title": "RiskSummary",
        "description": "Compact latest-risk view embedded in farm listings."
      },
      "ScenarioOut": {
        "properties": {
          "scenario": {
            "type": "string",
            "enum": [
              "normal",
              "dry_spell",
              "heavy_rain"
            ],
            "title": "Scenario",
            "description": "Active mock weather scenario"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "Active weather provider (mock | conduit)"
          },
          "available": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available"
          }
        },
        "type": "object",
        "required": [
          "scenario",
          "provider",
          "available"
        ],
        "title": "ScenarioOut"
      },
      "ScenarioSwitchOut": {
        "properties": {
          "scenario": {
            "type": "string",
            "enum": [
              "normal",
              "dry_spell",
              "heavy_rain"
            ],
            "title": "Scenario"
          },
          "reassessed": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RiskSummary"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Reassessed",
            "description": "New summaries per farm (if reassess=true)"
          },
          "alerts_sent": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Alerts Sent",
            "description": "Alert ids sent when notify=true"
          }
        },
        "type": "object",
        "required": [
          "scenario"
        ],
        "title": "ScenarioSwitchOut"
      },
      "ScenarioUpdate": {
        "properties": {
          "scenario": {
            "type": "string",
            "enum": [
              "normal",
              "dry_spell",
              "heavy_rain"
            ],
            "title": "Scenario"
          },
          "reassess": {
            "type": "boolean",
            "title": "Reassess",
            "description": "Re-run the risk engine for every farm after switching",
            "default": true
          },
          "notify": {
            "type": "boolean",
            "title": "Notify",
            "description": "After re-assessing, send SMS alerts to farmers whose risk change warrants one (dedupe rules apply)",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "scenario"
        ],
        "title": "ScenarioUpdate",
        "example": {
          "reassess": true,
          "scenario": "dry_spell"
        }
      },
      "StageOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Growth stage, e.g. flowering"
          },
          "day_after_planting": {
            "type": "integer",
            "title": "Day After Planting"
          },
          "day_in_stage": {
            "type": "integer",
            "title": "Day In Stage"
          },
          "stage_length_days": {
            "type": "integer",
            "title": "Stage Length Days"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "description": "0-1 progress through the season"
          },
          "water_need_mm_week": {
            "type": "number",
            "title": "Water Need Mm Week",
            "description": "FAO-56 crop water requirement for this stage"
          },
          "sensitivity": {
            "type": "number",
            "title": "Sensitivity",
            "description": "0-1 yield sensitivity to stress in this stage"
          },
          "is_critical": {
            "type": "boolean",
            "title": "Is Critical"
          }
        },
        "type": "object",
        "required": [
          "name",
          "day_after_planting",
          "day_in_stage",
          "stage_length_days",
          "progress",
          "water_need_mm_week",
          "sensitivity",
          "is_critical"
        ],
        "title": "StageOut"
      },
...
      "SubScoresOut": {
        "properties": {
          "drought": {
            "$ref": "#/components/schemas/SubScoreOut"
          },
          "flood": {
            "$ref": "#/components/schemas/SubScoreOut"
          },
          "heat": {
            "$ref": "#/components/schemas/SubScoreOut"
          },
          "crop_health": {
            "$ref": "#/components/schemas/SubScoreOut"
          }
        },
        "type": "object",
        "required": [
          "drought",
          "flood",
          "heat",
          "crop_health"
        ],
        "title": "SubScoresOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "PartnerApiKey": {
        "type": "apiKey",
        "description": "Partner API key issued to insurers, banks, SACCOs and agribusinesses. Demo keys: `fs_demo_acme_insurance_2026` (acme-insurance), `fs_demo_harvest_sacco_2026` (harvest-sacco).",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}